# public/document/api-doc.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample web server for Smart-City project.",
    "version": "1.0.0",
    "title": "Smart-City Backend API Doc"
  },
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/admin/data/list": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "getAllData",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/admin/pending-data/list": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "getPendingData",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/admin/user/list": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "getAllUsers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/admin/pending-user/list": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "getPendingUsers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/admin/pending-publish/list": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "getPendingPublish",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/admin/data/validate": {
      "put": {
        "tags": [
          "Admin"
        ],
        "operationId": "ValidateData",
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fileName",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/admin/user/validate": {
      "put": {
        "tags": [
          "Admin"
        ],
        "operationId": "ValidateUser",
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/admin/publish/validate": {
      "put": {
        "tags": [
          "Admin"
        ],
        "operationId": "ValidatePublish",
        "consumes": [
          "application/json",
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "publishId",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "shapefiles",
            "in": "formData",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/admin/data/cancel": {
      "put": {
        "tags": [
          "Admin"
        ],
        "operationId": "cancelDataApplication",
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fileName",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/admin/user/cancel": {
      "put": {
        "tags": [
          "Admin"
        ],
        "operationId": "cancelUserApplication",
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/admin/publish/cancel": {
      "put": {
        "tags": [
          "Admin"
        ],
        "operationId": "cancelPublishApplication",
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "publishId",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
...
    "/api/publish": {
      "post": {
        "tags": [
          "Publish"
        ],
        "operationId": "publishService",
        "consumes": [
          "application/json",
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "arcgisUser",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "arcgisPassword",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "arcgisPortal",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "serviceName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "fileNames",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "shapefiles",
            "in": "formData",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
...
  }
}